+++ src/main/java/ifpb/app_sistema_gestao_eventos/service/InscricaoService.java
@@ -1,6 +1,8 @@
 package ifpb.app_sistema_gestao_eventos.service;
 
+import ifpb.app_sistema_gestao_eventos.exception.EntidadeJaCadastradaException;
 import ifpb.app_sistema_gestao_eventos.exception.EntidadeNaoEncontradaException;
+import ifpb.app_sistema_gestao_eventos.exception.RegraDeNegocioException;
 import ifpb.app_sistema_gestao_eventos.mapper.InscricaoMapper;
 import ifpb.app_sistema_gestao_eventos.model.dto.InscricaoRequestDTO;
 import ifpb.app_sistema_gestao_eventos.model.dto.InscricaoResponseDTO;
@@ -10,9 +12,9 @@ import ifpb.app_sistema_gestao_eventos.model.entity.Usuario;
 import ifpb.app_sistema_gestao_eventos.repository.EventoRepository;
 import ifpb.app_sistema_gestao_eventos.repository.InscricaoRepository;
 import ifpb.app_sistema_gestao_eventos.repository.UsuarioRepository;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 import static ifpb.app_sistema_gestao_eventos.mapper.InscricaoMapper.toInscricaoResponseDTO;
 
@@ -29,11 +31,9 @@ public class InscricaoService {
         this.eventoRepository = eventoRepository;
     }
 
-    public List<InscricaoResponseDTO> listarInscricoes() {
-        return inscricaoRepository.findAll()
-                .stream()
-                .map(InscricaoMapper::toInscricaoResponseDTO)
-                .toList();
+    public Page<InscricaoResponseDTO> listarInscricoes(Pageable pageable) {
+        return inscricaoRepository.findAll(pageable)
+                .map(InscricaoMapper::toInscricaoResponseDTO);
     }
 
     public InscricaoResponseDTO buscarInscricaoPorId(Long id) {
@@ -43,20 +43,34 @@ public class InscricaoService {
     }
 
     public InscricaoResponseDTO salvarInscricao(InscricaoRequestDTO inscricao) {
+        if (inscricaoRepository.existsByUsuarioIdAndEventoId(inscricao.usuarioId(), inscricao.eventoId())) {
+            throw new EntidadeJaCadastradaException("Usuário já inscrito neste evento");
+        }
         Usuario usuario = usuarioRepository.findById(inscricao.usuarioId()).orElseThrow(() -> new EntidadeNaoEncontradaException("Usuário não encontrado"));
         Evento evento = eventoRepository.findById(inscricao.eventoId()).orElseThrow(() -> new EntidadeNaoEncontradaException("Evento não encontrado"));
+        if (inscricaoRepository.countByEventoId(evento.getId()) >= evento.getSala().getCapacidade()) {
+            throw new RegraDeNegocioException("Evento sem vagas disponíveis");
+        }
         Inscricao novaInscricao = new Inscricao(usuario, evento);
-        inscricaoRepository.save(novaInscricao);
-        return toInscricaoResponseDTO(novaInscricao);
+        return toInscricaoResponseDTO(inscricaoRepository.save(novaInscricao));
     }
 
     public InscricaoResponseDTO atualizarInscricao(Long id, InscricaoRequestDTO inscricao) {
         Inscricao inscricaoAtualizada = inscricaoRepository.findById(id)
                 .orElseThrow(() -> new EntidadeNaoEncontradaException("Inscrição não encontrada"));
+        Usuario usuario = usuarioRepository.findById(inscricao.usuarioId())
+                .orElseThrow(() -> new EntidadeNaoEncontradaException("Usuário não encontrado"));
+        Evento evento = eventoRepository.findById(inscricao.eventoId())
+                .orElseThrow(() -> new EntidadeNaoEncontradaException("Evento não encontrado"));
+        inscricaoAtualizada.setUsuario(usuario);
+        inscricaoAtualizada.setEvento(evento);
         return InscricaoMapper.toInscricaoResponseDTO(inscricaoRepository.save(inscricaoAtualizada));
     }
 
     public void deletarInscricao(Long id) {
+        if (!inscricaoRepository.existsById(id)) {
+            throw new EntidadeNaoEncontradaException("Inscrição não encontrada");
+        }
         inscricaoRepository.deleteById(id);
     }
 }
